Move eligibility-warnings component to blocks
