+++ assets/stylesheets/_components.scss
@@ -21,6 +21,7 @@
 @import 'blocks/domain-to-plan-nudge/style';
 @import 'blocks/edit-gravatar/style';
 @import 'blocks/follow-button/style';
+@import 'blocks/eligibility-warnings/style';
 @import 'blocks/image-editor/style';
 @import 'blocks/like-button/style';
 @import 'blocks/plan-thank-you-card/style';
@@ -62,7 +63,6 @@
 @import 'components/domains/map-domain-step/style';
 @import 'components/domains/register-domain-step/style';
 @import 'components/drop-zone/style';
-@import 'components/eligibility-warnings/style';
 @import 'components/ellipsis-menu/style';
 @import 'components/email-verification/style';
 @import 'components/emojify/style';
