+++ client/blocks/eligibility-warnings/style.scss
@@ -7,6 +7,7 @@
 
 .eligibility-warnings .card {
     margin: 0;
+    display: flex;
 }
 
 .eligibility-warnings__message {
